Proved betaV-Trm: reordering substitutions by subst-suc & lemma34.
The goal wasn't in the right form: subst \rho was on the outside
and needed to be pushed in before appeal to ->\betaV-step. Using
subst-equiv, subst-suc and lemma34 we can reorder the substitution
to push \rho inside and the < V /var\0> outside.

lemma35 and lemma34 are "dual" in the sense that they take p `. U
to subst (E [U]) p and (subst E p) [U], respectively.